API_app/requirements.txt
fastapi
uvicorn
python-multipart
confluent-kafka==2.3.0
pymongo==4.6.1
Pillow==10.2.0
pydantic-settings==2.2.1
